requirements.txt
asgiref==3.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
confusable-homoglyphs==3.2.0
cryptography==37.0.1
defusedxml==0.7.1
Django==4.0.4
django-filter==21.1
django-model-utils==4.2.0
django-widget-tweaks==1.4.12
idna==3.3
mysqlclient==2.1.0
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
swapper==1.3.0
tzdata==2022.1
urllib3==1.26.9
